rework the autocompletion
Now we accept a list from an autocomplete list handler(_acl_hdlr).
This list will be shown to a user as is. No processing is applied to it.
Autocompletion routine detect a completion suffix(common prefix of every
tokens from the list returned by a handler) which must be added to a
uncompleted token(a token that have a cursor right after it).
It is not necessarily uncompleted token, but we think so. If this
uncompleted token is a prefix of a common prefix for the list, then
we remove this chars from a common prefix and the result string
will be a detected completion suffix.